Finally got minted working in beamer with fragile
Fragile option required according to minted docs: "Due to how Beamer treats verbatim
content, you may need to use either the fragile or fragile=singleslide
options for frames that contain minted commands and environments.
fragile=singleslide works best, but it disables overlays. fragile works
by saving the contents of each frame to a temp file and then reusing them. This
approach allows overlays, but will break if you have the string \end{frame}
at the beginning of a line (for example, in a minted environment). To work
around that, you can indent the content of the environment (so that the
\end{frame} is preceded by one or more spaces) and then use the gobble
or autogobble options to remove the indentation." http://tug.ctan.org/macros/latex/contrib/minted/minted.pdf